==> Code/01_select_data_to_label.ipynb <==
# Select data to label

import pandas as pd
import numpy as np

# Import data from pickle
paragraph_df = pd.read_pickle('../Data/paragraph_df.pkl')
sc_only_para_df = paras_no_sents_df[paras_no_sents_df['Supply_Chain']=='Yes']

sc_only_para_df.head(3)

percent_10 = len(sc_only_para_df) *.1
print(percent_10)

# Send to Excel
sampled_df = sc_only_para_df.sample(round(percent_10))
sampled_df.to_excel('../Data/sampled_paragraphs.xlsx')
