==> Code/02_train-test-val_split.ipynb <==
# Train-test-validation split

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

# Import labelled data from Excel
labelled_data = pd.read_excel('../Data/labelled_data.xlsx')

# Ensure datatypes are correct
labelled_data['Paragraph'] = labelled_data['Paragraph'].astype(str)
labelled_data['Process_action'] = labelled_data['Process_action'].astype('int')
labelled_data['Market_action'] = labelled_data['Market_action'].astype('int')

# Save labelled data to pickle for future use
final_label_set.to_pickle('../Data/final_label_set.pkl')

# Set index and check
labelled_data.set_index('Index',inplace=True)
labelled_data.head(2)

# Create a column that combines labels in all four classification tasks (to use with split)
labelled_data['All4'] = labelled_data.apply(lambda x: str(x['Process_action'])+str(x['Market_action'])+str(x['Social'])+str(x['Environment']), 
                                                    axis=1)
labelled_data['All4'].value_counts()

# Split into train, val, and test using 80-10-10 split
train, val_test = train_test_split(labelled_data, 
                                   test_size=0.2, 
                                   random_state=42, 
                                   stratify=labelled_data['All4'])
val, test = train_test_split(val_test,  
                             test_size=0.5, 
                             random_state=42, 
                             stratify=val_test['All4'])

train.to_pickle('../Data/train.pkl')
val.to_pickle('../Data/val.pkl')
test.to_pickle('../Data/test.pkl')
