==> Code/09_alt_predictions_from_SVM.ipynb <==
# Get SVM predictions

# Import modules
import pandas as pd
from sklearn.pipeline import Pipeline, make_pipeline
from sklearn.metrics import confusion_matrix, accuracy_score, classification_report, roc_auc_score, f1_score
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.svm import SVC
from sklearn.utils import shuffle

# Load saved baseline models
with open("../Data/baseline_models.txt", "rb") as f:
    baseline_models = pickle.load(f)

aug_svm_dict = baseline_models['aug_svm_dict']
svm_dict = baseline_models['svm_dict']

# Get alternate predictions using SVM (& save)
to_pred = pd.read_pickle('../Data/paragraph_df.pkl')
to_pred = to_pred[to_pred['Supply_Chain']=='Yes']

to_pred['Process_action_alt'] = aug_svm_dict['Process_action']['Model'].predict(to_pred['Paragraph'])
to_pred['Social_alt'] = svm_dict['Social']['Model'].predict(to_pred['Paragraph'])

to_pred.to_pickle('../Data/svm_alt_preds.pkl')

to_pred.head()
