==> Code/14_stats_kruskal-wallis.ipynb <==
# Industry differences statistical test

This notebook uses the Kruskal-Wallis test to test for differences in ESG supply chain content between different sub-sectors.

import pandas as pd
import numpy as np
import pickle
import statsmodels.formula.api as smf
import matplotlib.pyplot as plt
import seaborn as sns
from scipy import stats

# Import final dataframe
data = pd.read_pickle('../Data/Variable dataframes/company_details_pivoted.pkl')

### Hypothesis 1: Supply chain speaking space

#Get lists of SC paragraph share by industry
cons_dur_apparel = data[(data['main_industry']=='Consumer Durables and Apparel (Primary)')&(data['Share_SC'].notna())]['Share_SC'].to_list()
cons_services = data[(data['main_industry']=='Consumer Services (Primary)')&(data['Share_SC'].notna())]['Share_SC'].to_list()
retail = data[(data['main_industry']=='Retailing (Primary)')&(data['Share_SC'].notna())]['Share_SC'].to_list()
food_staples_retail = data[(data['main_industry']=='Food and Staples Retailing (Primary)')&(data['Share_SC'].notna())]['Share_SC'].to_list()
auto = data[(data['main_industry']=='Automobiles and Components (Primary)')&(data['Share_SC'].notna())]['Share_SC'].to_list()
food_bev_tob = data[(data['main_industry']=='Food, Beverage and Tobacco (Primary)')&(data['Share_SC'].notna())]['Share_SC'].to_list()
household_personal = data[(data['main_industry']=='Household and Personal Products (Primary)')&(data['Share_SC'].notna())]['Share_SC'].to_list()

# Check whether groups differ
stats.kruskal(cons_dur_apparel, cons_services, retail, 
              food_staples_retail, auto, food_bev_tob, household_personal)

# Need to check which groups are different

ind_lists = [cons_dur_apparel, cons_services, retail, 
             food_staples_retail, auto, food_bev_tob, household_personal]
names = ['Consumer Durables and Apparel','Consumer Services','Retailing',
         'Food and Staples Retailing','Automobiles and Components',
         'Food, Beverage and Tobacco','Household and Personal Products']

kruskal_dict = dict()

for ind1, name1 in zip(ind_lists, names):
    kruskal_dict[name1] = dict()
    for ind2, name2 in zip(ind_lists, names):
        stat, p = stats.kruskal(ind1, ind2)
        kruskal_dict[name1][name2] = f"{stat:.2f} ({p:.3f})"
        
kruskal_df = pd.DataFrame(kruskal_dict)
display(kruskal_df)

### Hypothesis 2: Process-to-market speaking space index

#Get lists of process-to-market by industry
cons_dur_apparel = data[(data['main_industry']=='Consumer Durables and Apparel (Primary)')&(data['Market_to_process_log'].notna())]['Market_to_process_log'].to_list()
cons_services = data[(data['main_industry']=='Consumer Services (Primary)')&(data['Market_to_process_log'].notna())]['Market_to_process_log'].to_list()
retail = data[(data['main_industry']=='Retailing (Primary)')&(data['Market_to_process_log'].notna())]['Market_to_process_log'].to_list()
food_staples_retail = data[(data['main_industry']=='Food and Staples Retailing (Primary)')&(data['Market_to_process_log'].notna())]['Market_to_process_log'].to_list()
auto = data[(data['main_industry']=='Automobiles and Components (Primary)')&(data['Market_to_process_log'].notna())]['Market_to_process_log'].to_list()
food_bev_tob = data[(data['main_industry']=='Food, Beverage and Tobacco (Primary)')&(data['Market_to_process_log'].notna())]['Market_to_process_log'].to_list()
household_personal = data[(data['main_industry']=='Household and Personal Products (Primary)')&(data['Market_to_process_log'].notna())]['Market_to_process_log'].to_list()

# Check whether groups differ
stats.kruskal(cons_dur_apparel, cons_services, retail, 
              food_staples_retail, auto, food_bev_tob, household_personal)

# Need to check which groups are different

ind_lists = [cons_dur_apparel, cons_services, retail, 
             food_staples_retail, auto, food_bev_tob, household_personal]
names = ['Consumer Durables and Apparel','Consumer Services','Retailing',
         'Food and Staples Retailing','Automobiles and Components',
         'Food, Beverage and Tobacco','Household and Personal Products']

kruskal_dict = dict()

for ind1, name1 in zip(ind_lists, names):
    kruskal_dict[name1] = dict()
    for ind2, name2 in zip(ind_lists, names):
        stat, p = stats.kruskal(ind1, ind2)
        kruskal_dict[name1][name2] = f"{stat:.2f} ({p:.3f})"
        
kruskal_df = pd.DataFrame(kruskal_dict)
display(kruskal_df)

### Hypothesis 3: Environment-to-social speaking space index

#Get lists of SC paragraph share by industry
cons_dur_apparel = data[(data['main_industry']=='Consumer Durables and Apparel (Primary)')&(data['Env_to_social_log'].notna())]['Env_to_social_log'].to_list()
cons_services = data[(data['main_industry']=='Consumer Services (Primary)')&(data['Env_to_social_log'].notna())]['Env_to_social_log'].to_list()
retail = data[(data['main_industry']=='Retailing (Primary)')&(data['Env_to_social_log'].notna())]['Env_to_social_log'].to_list()
food_staples_retail = data[(data['main_industry']=='Food and Staples Retailing (Primary)')&(data['Env_to_social_log'].notna())]['Env_to_social_log'].to_list()
auto = data[(data['main_industry']=='Automobiles and Components (Primary)')&(data['Env_to_social_log'].notna())]['Env_to_social_log'].to_list()
food_bev_tob = data[(data['main_industry']=='Food, Beverage and Tobacco (Primary)')&(data['Env_to_social_log'].notna())]['Env_to_social_log'].to_list()
household_personal = data[(data['main_industry']=='Household and Personal Products (Primary)')&(data['Env_to_social_log'].notna())]['Env_to_social_log'].to_list()

# Check whether groups differ
stats.kruskal(cons_dur_apparel, cons_services, retail, 
              food_staples_retail, auto, food_bev_tob, household_personal)

# Need to check which groups are different

ind_lists = [cons_dur_apparel, cons_services, retail, 
             food_staples_retail, auto, food_bev_tob, household_personal]
names = ['Consumer Durables and Apparel','Consumer Services','Retailing',
         'Food and Staples Retailing','Automobiles and Components',
         'Food, Beverage and Tobacco','Household and Personal Products']

kruskal_dict = dict()

for ind1, name1 in zip(ind_lists, names):
    kruskal_dict[name1] = dict()
    for ind2, name2 in zip(ind_lists, names):
        stat, p = stats.kruskal(ind1, ind2)
        kruskal_dict[name1][name2] = f"{stat:.2f} ({p:.3f})"
        
kruskal_df = pd.DataFrame(kruskal_dict)
display(kruskal_df)

### Hypothesis 4: Supply chain topic breakdown

topic_list = ['Management systems_percent','Deforestation_percent',
                       'Human rights_percent','Employee health & safety_percent',
                       'Resource usage_percent',
                       'Certifications & training_percent',
                       'Collaborations & partnerships_percent',
                       'Plans and progress_percent',
                       'Governance & stakeholders_percent','Policies_percent',
                       'Product quality_percent',
                       'Diversity & inclusion_percent','Agriculture_percent',
                       'Risk assessments_percent','Chemicals_percent',
                       'Transportation & logistics_percent',
                       'Society_percent','Store operations_percent',
                       'Materials & packaging_percent']

topic_kruskal_dict = dict()

for topic in topic_list:
    #Get lists of SC paragraph share by industry
    cons_dur_apparel = data[data['main_industry']=='Consumer Durables and Apparel (Primary)'][topic].to_list()
    cons_services = data[data['main_industry']=='Consumer Services (Primary)'][topic].to_list()
    retail = data[data['main_industry']=='Retailing (Primary)'][topic].to_list()
    food_staples_retail = data[data['main_industry']=='Food and Staples Retailing (Primary)'][topic].to_list()
    auto = data[data['main_industry']=='Automobiles and Components (Primary)'][topic].to_list()
    food_bev_tob = data[data['main_industry']=='Food, Beverage and Tobacco (Primary)'][topic].to_list()
    household_personal = data[data['main_industry']=='Household and Personal Products (Primary)'][topic].to_list()
    
    # Get stats
    stat, p = stats.kruskal(cons_dur_apparel, cons_services, retail, 
              food_staples_retail, auto, food_bev_tob, household_personal, nan_policy='omit')
    
    topic_kruskal_dict[topic] = f"{stat:.2f} ({p:.3f})"
    
topic_kruskal_dict

pd.Series(topic_kruskal_dict)
